requirements.txt
django
matplotlib
numpy
pillow
scipy
tensorflow
object_detection
